requirements.txt
torch>=2.7.0
torchvision>=0.22.0
torchaudio>=2.7.0
transformers>=4.57.0
datasets>=4.4.0
pillow>=12.0.0
numpy>=2.3.0
huggingface-hub>=0.36.0
pyyaml>=6.0.3
tqdm>=4.67.0
accelerate>=1.12.0
peft>=0.18.0
trl>=0.26.0
einops>=0.8.0
safetensors>=0.7.0
tokenizers>=0.22.0
pandas>=2.3.0
matplotlib>=3.10.0
jupyter>=1.0.0
ipykernel>=7.1.0
ipywidgets>=8.1.0